pyproject.toml
[project]
name = "mlutils"
version = "0.1.0"
description = ""
authors = [
    {name = "Varshit Dusad",email = "dusadvarshit@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "pydantic (>=2.11.5,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "mlflow (>=2.22.0,<3.0.0)",
    "xgboost (>=3.0.1,<4.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "optuna (>=4.3.0,<5.0.0)",
    "optuna-integration (>=4.3.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "evidently (>=0.7.5,<0.8.0)",
    "imbalanced-learn (>=0.13.0,<0.14.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "boto3 (>=1.38.32,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
ruff = "^0.11.10"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
pydantic = "^2.11.4"

[tool.ruff]
line-length = 300

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
